+++ staff/francis-hrom/course-java/salvo-api/settings.gradle
@@ -1,0 +1,1 @@
+rootProject.name = 'salvo-api'
