+++ staff/francis-hrom/course-java/salvo-api/src/main/java/com/codeoftheweb/salvo_api/Game.java
@@ -1,12 +1,10 @@
-package com.codeoftheweb.salvo;
+package com.codeoftheweb.salvo_api;
 
 import org.hibernate.annotations.GenericGenerator;
 
 import javax.persistence.*;
 
 import java.util.Date;
-import java.util.LinkedHashMap;
-import java.util.Map;
 import java.util.Set;
