+++ staff/francis-hrom/course-java/salvo-api/src/main/java/com/codeoftheweb/salvo_api/GamePlayer.java
@@ -1,4 +1,4 @@
-package com.codeoftheweb.salvo;
+package com.codeoftheweb.salvo_api;
 
 import org.hibernate.annotations.GenericGenerator;
 
@@ -7,15 +7,8 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import java.util.Date;
-import java.util.LinkedHashMap;
-import java.util.Map;
 
-import org.hibernate.annotations.GenericGenerator;
-import javax.persistence.Entity;
 import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
