+++ staff/francis-hrom/course-java/salvo-api/src/main/java/com/codeoftheweb/salvo_api/GamePlayerRepository.java
@@ -1,4 +1,4 @@
-package com.codeoftheweb.salvo;
+package com.codeoftheweb.salvo_api;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
