+++ staff/francis-hrom/course-java/salvo-api/src/main/java/com/codeoftheweb/salvo_api/GameRepository.java
@@ -1,8 +1,8 @@
-package com.codeoftheweb.salvo;
+package com.codeoftheweb.salvo_api;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource
-public interface PlayerRepository extends JpaRepository<Player, Long> {
+public interface GameRepository extends JpaRepository<Game, Long> {
 }
