+++ staff/francis-hrom/course-java/salvo-api/src/main/java/com/codeoftheweb/salvo_api/Player.java
@@ -1,4 +1,4 @@
-package com.codeoftheweb.salvo;
+package com.codeoftheweb.salvo_api;
 
 import org.hibernate.annotations.GenericGenerator;
 import javax.persistence.Entity;
@@ -6,21 +6,10 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 
-import java.util.Date;
-import org.hibernate.annotations.GenericGenerator;
-import javax.persistence.Entity;
 import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
 
 import javax.persistence.OneToMany;
 import java.util.Set;
-import java.util.HashSet;
-
-import java.util.List;
 
 
 @Entity
