+++ staff/francis-hrom/course-java/salvo-api/src/main/resources/static/manager.html
