+++ staff/francis-hrom/course-java/salvo-api/src/test/java/com/codeoftheweb/salvo_api/SalvoApplicationTests.java
@@ -1,4 +1,4 @@
-package com.codeoftheweb.salvo;
+package com.codeoftheweb.salvo_api;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
