--- staff/francis-hrom/course-java/salvo/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = 'salvo'
